+++ CHANGELOG.md
@@ -1,3 +1,8 @@
-## 1.0.0
+## 0.0.2-beta
 
-- Initial version.
+- Included default value for getValue
+- If path is null or empty, getValue will return all json contents
+
+## 0.0.1-beta
+
+- Initial version created by GnRDevAnd Team